PLANNER-454 HardSoftBigDecimalScore and SimpleBigDecimalScore's toLevelNumbers() should return BigDecimals, not Doubles.

+++ optaplanner-core/src/main/java/org/optaplanner/core/api/score/buildin/hardsoftbigdecimal/HardSoftBigDecimalScore.java
@@ -136,7 +136,7 @@ public final class HardSoftBigDecimalScore extends AbstractScore<HardSoftBigDeci
     }
 
     public Number[] toLevelNumbers() {
-        return new Number[]{hardScore.doubleValue(), softScore.doubleValue()};
+        return new Number[]{hardScore, softScore};
     }
 
     public boolean equals(Object o) {
